Acces method started, modified Process so that it stores the id of the page in the array, modified the changeProcess method

+++ Time.java
@@ -20,7 +20,7 @@ public class Time {
 	}
 
 	// Overwriting for different argument type
-	void addSeconds(float sec) {
+	void addSeconds(double sec) {
 		this.miliseconds += (sec % 1) * 1000;
 		this.seconds += sec - (sec % 1);
 		simplifyTime();
